Change IntSet::len() to return u64 instead of usize.
A u32 containing intset can have up to u32::MAX + 1 elements in it which can overflow a usize when usize is a u32.

+++ fuzz/fuzz_targets/fuzz_sparse_bit_set_encode.rs
@@ -8,7 +8,7 @@ mod int_set_op_processor;
 use int_set_op_processor::process_op_codes;
 use int_set_op_processor::OperationSet;
 
-const OPERATION_COUNT: usize = 2000;
+const OPERATION_COUNT: u64 = 2000;
 
 fuzz_target!(|data: &[u8]| {
     let _ = process_op_codes::<u32>(OperationSet::SparseBitSetEncoding, OPERATION_COUNT, data);
